requirements.txt
PyQt5
opencv-python
pandas
numpy
boto3
avro
PyWavelets
scipy
py-feat
jsonschema
mss